requirements.txt
PyYaml
kubernetes
requests